move hydra-zen config from repro to library
